refactor(vscode): extract the rbx spawn and discovery helper
